# docs/openapi.yaml
openapi: 3.1.0
info:
  title: API Carte UAC
  version: 1.0.0
  description: |
    API de la carte du campus de l'Université d'Abomey-Calavi.
    Toutes les réponses JSON suivent la même enveloppe : `{ data, meta? }` en cas de succès,
    `{ error: { code, message, status } }` en cas d'échec. Les listes sont toujours paginées (`meta`).
    Ce fichier est le contrat de référence : toute modification de l'API commence ici.
  license:
    name: MPL-2.0
    url: https://mozilla.org/MPL/2.0/
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: uac_admin_session
      description: Cookie httpOnly posé par POST /admin/session (appli web).
    adminToken:
      type: http
      scheme: bearer
      description: Jeton ADMIN_TOKEN, réservé aux scripts et à l'intégration continue.
    contributorCookie:
      type: apiKey
      in: cookie
      name: uac_contributor
      description: Cookie httpOnly posé par POST /contributors, valable 180 jours. Identifie un téléphone contributeur.
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Identifiant de traçage (8 à 64 caractères alphanumériques ou tirets). Généré s'il est absent, renvoyé dans la réponse.
      schema: { type: string }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    ReviewerId:
      name: reviewerId
      in: path
      required: true
      schema: { type: string }
    LinkId:
      name: linkId
      in: path
      required: true
      schema: { type: string }
  headers:
    RequestId:
      description: Identifiant de traçage de la requête.
      schema: { type: string }
  schemas:
...
    PaginationMeta:
      type: object
      required: [page, limit, total]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, status]
          properties:
            code: { type: string, example: PLACE_NOT_FOUND }
            message: { type: string, example: Lieu inexistant }
            status: { type: integer, example: 404 }
...
    Reviewer:
      type: object
      properties:
        id: { type: string }
        name: { type: string, maxLength: 80 }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        revokedAt: { type: [string, 'null'], format: date-time }
    ContributionLink:
      type: object
      properties:
        id: { type: string, description: "Code du lien, placé dans l'adresse ?contribuer=…" }
        label: { type: string, maxLength: 80 }
        isActive: { type: boolean }
        isPublic: { type: boolean, description: "Lien affiché sur le site ; au plus un lien actif l'est." }
        createdAt: { type: string, format: date-time }
        closedAt: { type: [string, 'null'], format: date-time }
        url: { type: string }
  responses:
    BadRequest:
      description: Données invalides (codes VALIDATION_ERROR, INVALID_JSON, INVALID_PAGE, INVALID_LIMIT, INVALID_SORT, INVALID_ORDER, INVALID_CATEGORY, INVALID_PATH_TYPE, INVALID_FORMAT, CONFIRMATION_REQUIRED, INVALID_STATUS, INVALID_ENTITY_TYPE, INVALID_CONTRIBUTOR_ID, INVALID_ENTITY_ID).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Session absente ou expirée (UNAUTHORIZED), ou jeton incorrect (INVALID_TOKEN).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Ressource inexistante (PLACE_NOT_FOUND, PATH_NOT_FOUND, ROUTE_NOT_FOUND, NO_PUBLIC_LINK, LINK_NOT_FOUND, TARGET_NOT_FOUND, PROPOSAL_NOT_FOUND, CONTRIBUTOR_NOT_FOUND, CHANGE_NOT_FOUND, REVIEWER_NOT_FOUND).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Téléphone bloqué (CONTRIBUTOR_BLOCKED), ou action réservée à l'administrateur (FORBIDDEN).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NoContent:
      description: Succès sans contenu.
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
security: []
paths:
...
  /reviewers:
    parameters: [{ $ref: '#/components/parameters/RequestId' }]
    get:
      tags: [Administration]
      summary: Liste des relecteurs
      security: [{ adminSession: [] }, { adminToken: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Page de relecteurs, du plus récent au plus ancien
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Reviewer' } }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Administration]
      summary: Créer un relecteur
      description: La réponse contient son jeton personnel, montré une seule fois.
      security: [{ adminSession: [] }, { adminToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 80 }
      responses:
        '201':
          description: Relecteur créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Reviewer'
                      - type: object
                        properties:
                          token: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reviewers/{reviewerId}:
    parameters: [{ $ref: '#/components/parameters/ReviewerId' }, { $ref: '#/components/parameters/RequestId' }]
    delete:
      tags: [Administration]
      summary: Révoquer un relecteur (ferme ses sessions)
      security: [{ adminSession: [] }, { adminToken: [] }]
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /contribution-links:
    parameters: [{ $ref: '#/components/parameters/RequestId' }]
    get:
      tags: [Administration]
      summary: Liste des liens de contribution
      security: [{ adminSession: [] }, { adminToken: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Page de liens, du plus récent au plus ancien
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/ContributionLink' } }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Administration]
      summary: Créer un lien de contribution
      security: [{ adminSession: [] }, { adminToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label: { type: string, maxLength: 80 }
                isPublic: { type: boolean, default: false }
      responses:
        '201':
          description: Lien créé
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/ContributionLink' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /contribution-links/{linkId}:
    parameters: [{ $ref: '#/components/parameters/LinkId' }, { $ref: '#/components/parameters/RequestId' }]
    patch:
      tags: [Administration]
      summary: Renommer, fermer, rouvrir un lien, ou le désigner comme lien public
      description: Fermer un lien lui retire le statut public ; un lien fermé ne peut pas être désigné public.
      security: [{ adminSession: [] }, { adminToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { type: string, maxLength: 80 }
                isActive: { type: boolean }
                isPublic: { type: boolean }
      responses:
        '200':
          description: Lien modifié
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/ContributionLink' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /campus-settings:
    parameters: [{ $ref: '#/components/parameters/RequestId' }]
    patch:
      tags: [Administration]
      summary: Suspendre ou reprendre les contributions
      security: [{ adminSession: [] }, { adminToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contributionsPaused]
              properties:
                contributionsPaused: { type: boolean }
      responses:
        '200':
          description: Réglages de la carte
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/CampusMap/properties/settings' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }